Add advanced hotkey support (Fn, modifier-only) and fix UInt32 crash
- Add CGEventTap backend for Fn key and modifier-only hotkeys (⌘ alone, etc.)
- Add Shortcut model with constants and conversion functions
- Fix crash when using modifier-only hotkeys (UInt32(-1) overflow)
- Improve hotkey recorder: accept normal keys on keyDown, modifier-only on release
- Remove main window, app is now menu bar only
- Add code signing to build script for Input Monitoring permission
- Add SettingsWindowController for proper settings window management
- Add 27 new hotkey recorder tests (67 total)

Known issue: Modifier-only activation feels slightly wonky, needs review.

+++ Sources/FlowDictate/Views/SettingsView.swift
@@ -79,8 +79,8 @@ private struct GeneralSettingsTab: View {
                     modifiers: $settingsManager.hotkeyModifiers,
                     onHotkeyChanged: {
                         appController.updateHotkey(
-                            keyCode: UInt32(settingsManager.hotkeyKeyCode),
-                            modifiers: UInt32(settingsManager.hotkeyModifiers)
+                            keyCode: settingsManager.hotkeyKeyCode,
+                            modifiers: settingsManager.hotkeyModifiers
                         )
                     }
                 )
